Dockerizing calculator service
In this commit we have made a "build/" folder for storing the Dockerfile
and necessary configurations.  We've added a uwsgi configuration file
and made a slight change to the calculator module so that
`calculator.app.app` can be found as `calculator.application`.

The second step is to be able to run tests inside the Docker.